+++ CHANGELOG.md
@@ -55,7 +55,8 @@ Example payload of the request to deferred-credential endpoint`/deferred_credent
 - Expanded `/token` functionality. The endpoint accepts now `application/x-www-urlencoded` and no content-type.
   It still accepts the values in the url as request-params (this functionality will be removed in the future) and
   as `x-www-form-urlencoded` body.
-
+- Optional OAuth security with bearer tokens on `/management` endpoints. 
+  It can be activated and configured via spring environment variables.
 ### Fixed
 
 - Checks for protected claims are now done in the create-offer-flow (1 step) instead of the issuance flow.